rm fixed dependency on ui-plugin-find-users 1.0.0
Apps should not depend explicitly on plugins. The WHOLE POINT of a
plugin is that it can be switched out at runtime. Also, the fact that
this was a hard dep on version 1.0.0 rather than, say, ^1.0.0, appears
to have caused multiple copies of the plugin and of its dependencies to
be built into the bundle and that caused no end of trouble.