use the "status" formatter for the "active" attr
`active` is the attribute on the user-object, but the formatter for this
field is named `status` because, um, I don't remember why but that's what
it's called now. So the `<Pluggable>` needs to ask for `status` rather
than `active` for that data to be correctly formatted in the modal.

Fixes UICHKOUT-472