use inputRef to access TextField
Shamelessly copied from @JohnC-80 in folio-org/ui-checkin#103

Checkin had to resort to pushing a ref through the HOC's that
wrap TextField in order to properly obtain a ref so that it
could call instance methods of TextField (`focusInput()`)
This isn't an ideal way for things to be shaped.

TextField has a prop: `inputRef` that can be used to obtain a
direct ref to the `<input>` rendered by TextField. This will
punch through any HOC's without needing to rely on
ref-forwarding or `withRef` API. Any usage of `focusInput()`
or `getInput()` can be replaced by the native `focus()`. In
this case, the API to dig up the rendered component
(`getRenderedComponent()`) can be removed as well.

Refs STCOM-394