UITEST-70 be less markupy, more semantic (#382)
Nightmare appears to be getting hung up in some MCL selections. I'm not
sure how this is coming up now given that this passed for several days,
but here we are. Removing some of the markup and leaving the selector to
stand on semantic markup alone seems to do the trick.

Refs UITEST-70