avoid admin users in tests
The code for plucking a user from the patron-look-up pop-up could allow
for diku_admin, a user without a barcode, to be selected. This caused
issues with downstream tests that expect the user to have a barcode.

Somebody utterly botched an earlier attempt to do this exact thing in
PR #404 but must have pushed up the wrong branch because that only
contained an extra timer, useful for inspecting what happened in the
test, but useless in terms of actually fixing it. What a moron. Another
nail in the coffin for having the privilege of merging your own PRs.
Sure, it can be expedient, but it can also be stupid. This is living
proof.