avoid accidentally selecting diku
`diku` is a user with an invalid user record (it's incomplete, missing
the `personal` object) that seems to be created by ... some unknown part
of somebody's process somewhere. It's a bit of a mystery. At any rate,
selecting this user causes all kinds of grief because the record is
invalid in ways that are normally protected against (i.e. if it's
impossible to create a record without a `personal` attribute, then we
don't have to check for its presence because such presence is
guaranteed) except here we are.