@folio devdeps MUST use caret deps
The dev-dep on `@folio/stripes-components` MUST be a caret dep
compatible with the selected version of stripes, otherwise it risks
pulling multiple copies into the build because of a newer version
present in `@folio/stripes`.

This bug was introduced in #441 and inadvertently carried through other
updates but we didn't realize it until now, when `v6.1.0` was published,
putting it out of sync with the `~6.0` version here.