UICHKOUT-614 restore checkout-as-proxy
`props.settings`, which is used to determine if a user's profile picture
should be displayed, but which we don't actually implement, was
inadvertently removed from the connected `UserDetail` component when
viewing a proxy, which caused an NPE, which made checkout-as-proxy
impossible to use.

IOW, a prop for a setting we _don't_ use broke a feature we _do_. Oy.

Restored the setting, and added some tests.

Refs UICHKOUT-614