UICHKOUT-639: Proxy borrower does not display on loan details
This turned out to be a problem with ui-checkout.  The module
was not providing the barcode of the proxy borrower when it
issued the check-out request, so as far as the back-end was
concerned, there was no proxy borrower.  I fixed this by
making the module check for the presence of proxy borrower
data, and include the proxy borrower's barcode if it was
present.

refs: https://issues.folio.org/browse/UICHKOUT-639