rename session-timeout test to run it last
I don't have a good explanation for why I thought this would work, but
it did. Here's the situation: tests were only seldom passing on master,
and were failing consistently on branches in PRs, but that same set of
tests _passed_ consistently when it was run with a `.only` in its
top-level `describe()` block.

I suspect there is some bad interaction with promises somewhere, but I'm
not sure where exactly. When the timeout test runs, there is zero output
until its timeout expires, then all the lines from its `describe` and
`it` blocks appear at once. That doesn't seem right, right? Anyway, here
we are.