UICHKOUT-646: Fast add permission (Check out)
This was very straightforward.  I just wrapped an
IfPermission around the button.  There are existing
checks happening elsewhere in the system that seem to
me to address the other requirements in this story,
such as ensuring only people with checkout permission
can access the form (they can't load the checkout
module at all anyway.)

Refs: https://issues.folio.org/browse/UICHKOUT-646