UICHKOUT-644 better proxy handling (#533)
* `fuggle`, verb, transitive: to screw something up
* `defuggle`, verb, transitive: to think you understand the problem when
in fact you do not, causing you to implement a fix that papers over the
issue without actually resolving it.
* `undefuggle`, verb, transitive: to truly understand the problem and
why the original "fix" didn't take, and to actually resolve the issue.

I don't know when this problem was originally fuggled, but the crux is
that the checkout API rejects requests that contain the same barcode for
both patron and proxy. #527 defuggled this by clearing out the proxy
value when a patron had sponsors but was acting as self. The hidden
problem was that the ProxyManager isn't smart enough to hide when
its patron and proxy values match; if a patron has sponsors and proxy is
empty, it _always_ displays.

In other words, `<ProxyManager>` and `<ScanItems>` have different needs
for how their proxy props are handled. *<sigh>* The final fix here is to
revert #527 and then update `<ScanItems>` to ignore proxy when patron
and proxy match.

Fuggle, defuggle, undefuggle. Now you know.

Refs UICHKOUT-644