UICHKOUT-513 restore item-list sort
It looks like the sort implementation got confused between column
aliases, which contain the displayed (i.e. i18n'ed) values, and column
names, which contain their code points. It's possible that before we had
correct i18n, the alias pointed to a simple string like `Barcode` rather
than a `<FormattedMessage id="barcode">` object that is meaningless
from a sorting point of view, as this was the original implementation
in PR #27.

Refs UICHKOUT-513