repair tests by waiting for the user to load
Most broken tests did not wait for the user pane to finish loading
before entering values in the item pane, but the item pane is configured
to error-out if the user values aren't already present. Thus, a little
awaiting went a long way to getting this suite largely back on track.

One test caused an NPE with an incomplete stub that was missing some
(now required) attributes.