UICHKOUT-855: Can add patron/staff info at checkout (#804)
Two new menu items are added to the Three-Dots menu shown with each checked-out item within a session. These are **Add patron info** and **Add staff info**. If either is selected, a model prompts the user to enter the note to be attached to the loan. Multiple notes may be attached of either the same or different kinds.

Fixes UICHKOUT-855.

## Approach

`src/ScanItems.js` has two new manifest entries: `loan` for a loan to be updated with PUT, and an associated stripes-connect local resource `loanId` to supply the ID for the `loan` resource. These are used by a new `addPatronOrStaffInfo` method, which is passed as a new prop to `<ViewItem>`.

`src/components/ViewItem/ViewItem.js` has two new state entries: `infoType` is initially undefined but can be set to `patronInfo` or `staffInfo` to cause the relevant modal to display; and `loanForInfo` is the loan to which the information is to be added. New buttons in the action menu set this state to pop up the modal.

`src/components/ViewItem/AddInfoDialog.js` provides this modal, which works in the standard way using elements from stripes-components. When a note is entered and the use clicks **Save & close**, the note is added using the `addPatronOrStaffInfo`, which has been passed into the model as a prop.

See https://issues.folio.org/browse/UICHKOUT-855