[UICHKOUT-967] Guard against multiple invocations of componentDidMount (#927)
When running in dev mode — and not in production — React 18 will deliberately run all `componentDidMount` methods twice. It does this to help you find places where you are inadvertently doing non-idempotent things, such as, well, checking out items. See https://react.dev/reference/react/Component#componentdidmount

The proper fix is to change the whole approach of ui-checkout so it doesn’t rely on side-effects of mounting a component. That would be a complicated job and I don’t understand the logic of the module well enough to confident I was doing that work right.

This commit simply adds guard code to the `componentDidMount` method so that it only runs once. That makes no difference to how the code runs in production, but makes it possible to do development.