[UICHKOUT-980] Reinstate display of ITEM_HAS_OPEN_LOAN error (#937)
When a user tries to check out an item that is already checked out, mod-circulation returns with HTTP status 422 Unprocessable Content, and a body that contains one or more error objects.

For reasons that are not clear, much of the time, it includes two errors which are more or less duplicates: "Cannot check out item that already has an open loan" with a `code` element of "ITEM_HAS_OPEN_LOAN", and "Item is already checked out".

UICHKOUT-695 fixed the double-reporting error by adding a `BACKEND_ERRORS_CODES_TO_HIDE` list which consisted only of `BACKEND_ERROR_CODES.itemHasOpenLoan`. (This is a constant equal to "ITEM_HAS_OPEN_LOAN"). The code then omits from the error modal any errors that have a `code` in that list.

But under some circumstance — for example, on nls-dev — the "Item is already checked out" version of the error is not included in mod-circulation's response. As a result, the only included error message is discarded, and the error popup says _only_ "Item not checked out" with no explanation of why not.

This commit trivially modifies the `BACKEND_ERRORS_CODES_TO_HIDE` code so that it will not omit an error whose code is listed if it's the only one in the response. This means that on systems (such as snapshot) where both errors are included, the behaviour is unchanged — the duplicate error is still removed. But on systems (like nls-dev) where only one error is included in the response, that is displayed.